Fix ostree_* upgrade
Currently, the code assumes that both
ostree_commit and ostree_remote are present
in cluster_template.labels. If one of them are
missing, the ostree upgrade fails [1] and leaves
the cluster with UPDATE_FAILED status.

By the docs[2], it is understood that users have
the ability to choose only one of the labels.

[1] https://gist.github.com/guilhermesteinmuller/7bf9f51e421283783cf737900797232c
[2] https://github.com/openstack/magnum/blob/master/doc/source/user/rolling-upgrade.rst

Change-Id: I0f65169305ba74c082b65bf39083def278404b93